added relu activation

+++ internal/transforms/transforms.go
@@ -4,8 +4,8 @@ import (
 	t "gorgonia.org/tensor"
 )
 
-func Bias(newShape t.Shape, values []float32) *t.Dense {
-	newBacking := make([]float32, newShape[0]*newShape[1])
+func Bias(newShape t.Shape, values []float64) *t.Dense {
+	newBacking := make([]float64, newShape[0]*newShape[1])
 	for i := range newBacking {
 		newBacking[i] = values[i%len(values)]
 	}
